refactor: reorganize OS files
Ref eng-268
