+++ packages/api/src/external/commonwell/document/document-query.ts
@@ -35,7 +35,7 @@ import { getPatientOrFail } from "../../../command/medical/patient/get-patient";
 import { Config } from "../../../shared/config";
 import { mapDocRefToMetriport } from "../../../shared/external";
 import { reportMetric } from "../../aws/cloudwatch";
-import { ingestIntoSearchEngine } from "@metriport/core/external/opensearch/file/ingest";
+import { ingestIntoSearchEngine } from "@metriport/core/command/consolidated/search/document-reference/ingest";
 import { convertCDAToFHIR, isConvertible } from "../../fhir-converter/converter";
 import { makeFhirApi } from "../../fhir/api/api-factory";
 import { cwToFHIR } from "../../fhir/document";
