+++ packages/core/src/command/consolidated/search/fhir-resource/search-semantic.ts
@@ -1,14 +1,14 @@
 import { SearchSetBundle } from "@metriport/shared/medical";
-import { getConsolidatedPatientData } from "../consolidated-get";
-import { Patient } from "../../../domain/patient";
-import { out } from "../../../util";
-import { Config } from "../../../util/config";
-import { DocumentReferenceWithId } from "../../../external/fhir/document/document-reference";
-import { toFHIR as patientToFhir } from "../../../external/fhir/patient/conversion";
-import { buildBundleEntry } from "../../../external/fhir/shared/bundle";
-import { SearchResult } from "../../../external/opensearch/index-based-on-resource";
-import { searchDocuments } from "../../../external/opensearch/search-documents";
-import { OpenSearchSemanticSearcherDirect } from "../../../external/opensearch/semantic/semantic-searcher-direct";
+import { Patient } from "../../../../domain/patient";
+import { DocumentReferenceWithId } from "../../../../external/fhir/document/document-reference";
+import { toFHIR as patientToFhir } from "../../../../external/fhir/patient/conversion";
+import { buildBundleEntry } from "../../../../external/fhir/shared/bundle";
+import { SearchResult } from "../../../../external/opensearch/index-based-on-resource";
+import { OpenSearchSemanticSearcher } from "../../../../external/opensearch/semantic/semantic-searcher";
+import { out } from "../../../../util";
+import { Config } from "../../../../util/config";
+import { getConsolidatedPatientData } from "../../consolidated-get";
+import { searchDocuments } from "../document-reference/search";
 
 /**
  * Performs a semantic search on a patient's consolidated resources in OpenSearch
@@ -113,7 +113,7 @@ async function searchOpenSearch({
   const modelId = Config.getSemanticSearchModelId();
 
   // TODO eng-41 make this a factory so we can delegate the processing to a lambda
-  const searchService = new OpenSearchSemanticSearcherDirect({
+  const searchService = new OpenSearchSemanticSearcher({
     region,
     endpoint,
     indexName,
