--- packages/core/src/external/opensearch/document-status.ts
@@ -1,13 +1,0 @@
-import { DocumentReference } from "@medplum/fhirtypes";
-
-export function isDocStatusReady(doc: DocumentReference): boolean {
-  return !isDocStatusPreliminary(doc) && !isDocStatusEnteredInError(doc);
-}
-
-export function isDocStatusPreliminary(doc: DocumentReference): boolean {
-  return doc.docStatus === "preliminary";
-}
-
-export function isDocStatusEnteredInError(doc: DocumentReference): boolean {
-  return doc.docStatus === "entered-in-error";
-}
