+++ packages/core/src/external/opensearch/lexical/lexical-searcher.ts
@@ -2,15 +2,12 @@ import { BadRequestError } from "@metriport/shared";
 import { Client } from "@opensearch-project/opensearch";
 import { out } from "../../../util";
 import { OpenSearchConfigDirectAccess, OpenSearchResponse, OpenSearchResponseHit } from "../index";
-import { SearchResult } from "../index-based-on-resource";
-import { createHybridSearchQuery } from "./hybrid-search";
+import { indexDefinition, SearchResult } from "../index-based-on-resource";
+import { createLexicalSearchQuery } from "./lexical-search";
 
-const defaultSimilarityThreshold = 0.2;
 const defaultNumberOfResults = 100;
 
-export type OpenSearchSemanticSearcherDirectConfig = OpenSearchConfigDirectAccess & {
-  modelId: string;
-};
+export type OpenSearchLexicalSearcherConfig = OpenSearchConfigDirectAccess;
 
 export type SearchRequest = {
   cxId: string;
@@ -19,36 +16,30 @@ export type SearchRequest = {
   query: string;
   /** From 0 to 10_000, optional, defaults to 100 */
   maxNumberOfResults?: number | undefined;
-  /** From 0 to 1, optional, defaults to 0.2 */
-  similarityThreshold?: number | undefined;
 };
 
-export class OpenSearchSemanticSearcherDirect {
-  constructor(readonly config: OpenSearchSemanticSearcherDirectConfig) {}
+export class OpenSearchLexicalSearcher {
+  constructor(readonly config: OpenSearchLexicalSearcherConfig) {}
 
   async search({
     cxId,
     patientId,
     query,
     maxNumberOfResults = defaultNumberOfResults,
-    similarityThreshold = defaultSimilarityThreshold,
   }: SearchRequest): Promise<SearchResult[]> {
-    const { indexName, endpoint, username, password, modelId } = this.config;
-    const { log, debug } = out(`OpenSearchSemanticSearcherDirect - cx ${cxId}, pt ${patientId}`);
+    const { log, debug } = out(`${this.constructor.name}.search - cx ${cxId}, pt ${patientId}`);
 
     this.validateMaxNumberOfResults(maxNumberOfResults);
-    this.validateSimilarityThreshold(similarityThreshold);
 
+    const { indexName, endpoint, username, password } = this.config;
     const auth = { username, password };
     const client = new Client({ node: endpoint, auth });
 
     log(`Searching on index ${indexName}...`);
-    const queryPayload = createHybridSearchQuery({
+    const queryPayload = createLexicalSearchQuery({
       cxId,
       patientId,
       query,
-      modelId,
-      k: maxNumberOfResults,
     });
 
     const response = (
@@ -63,16 +54,7 @@ export class OpenSearchSemanticSearcherDirect {
     const items = response.hits.hits ?? [];
     log(`Successfully searched, got ${items.length} results`);
 
-    const filteredItems = items.filter(item => item._score >= similarityThreshold);
-    if (filteredItems.length !== items.length) {
-      log(
-        `Filtered ${
-          items.length - filteredItems.length
-        } results due to similarity threshold (${similarityThreshold})`
-      );
-    }
-
-    return this.mapResult(filteredItems);
+    return this.mapResult(items);
   }
 
   private validateMaxNumberOfResults(maxNumberOfResults: number) {
@@ -81,15 +63,6 @@ export class OpenSearchSemanticSearcherDirect {
     }
     if (maxNumberOfResults < 1) {
       throw new BadRequestError("maxNumberOfResults cannot be less than 1");
-    }
-  }
-
-  private validateSimilarityThreshold(similarityThreshold: number) {
-    if (similarityThreshold > 1) {
-      throw new BadRequestError("similarityThreshold cannot be greater than 1");
-    }
-    if (similarityThreshold <= 0) {
-      throw new BadRequestError("similarityThreshold cannot be less than or equal to 0");
     }
   }
 
@@ -103,5 +76,22 @@ export class OpenSearchSemanticSearcherDirect {
         resourceId: hit._source.resourceId,
       };
     });
+  }
+
+  async createIndexIfNotExists(): Promise<void> {
+    const { indexName, endpoint, username, password } = this.config;
+    const auth = { username, password };
+    const client = new Client({ node: endpoint, auth });
+
+    const indexExistsResp = await client.indices.exists({
+      index: indexName,
+      include_defaults: false,
+      ignore_unavailable: false,
+    });
+    const indexExists = Boolean(indexExistsResp.body);
+    if (indexExists) return;
+
+    const body = { mappings: { properties: indexDefinition } };
+    await client.indices.create({ index: indexName, body });
   }
 }
