+++ packages/core/src/external/opensearch/shared/bulk.ts
@@ -1,4 +1,6 @@
 import { ApiResponse } from "@opensearch-project/opensearch";
+import { IngestRequest } from "../text-ingestor";
+import { IngestRequestResource } from "../text-ingestor";
 
 export type BulkOperation = "index" | "create" | "update" | "delete";
 
@@ -63,4 +65,29 @@ export function bulkResponseErrorToString(
   const errorAsAny = itemOp.error?.reason;
   if (!errorAsAny) return undefined;
   return typeof errorAsAny === "string" ? errorAsAny : JSON.stringify(errorAsAny);
+}
+
+export function resourceToBulkRequest({
+  cxId,
+  patientId,
+  resource,
+  operation,
+  getEntryId,
+}: {
+  cxId: string;
+  patientId: string;
+  resource: IngestRequestResource;
+  operation: BulkOperation;
+  getEntryId: (cxId: string, patientId: string, resourceId: string) => string;
+}) {
+  const entryId = getEntryId(cxId, patientId, resource.resourceId);
+  const document: IngestRequest = {
+    cxId,
+    patientId,
+    resourceType: resource.resourceType,
+    resourceId: resource.resourceId,
+    content: resource.content,
+  };
+  const cmd = { [operation]: { _id: entryId } };
+  return [cmd, document];
 }
