+++ packages/core/src/external/opensearch/shared/filters.ts
@@ -1,0 +1,3 @@
+export function getPatientFilters(cxId: string, patientId: string) {
+  return [{ term: { cxId } }, { term: { patientId } }];
+}
