+++ packages/core/src/external/opensearch/text-ingestor.ts
@@ -1,19 +1,19 @@
-import { emptyFunction } from "@metriport/shared";
 import { Client } from "@opensearch-project/opensearch";
 import dayjs from "dayjs";
 import duration from "dayjs/plugin/duration";
 import { chunk } from "lodash";
 import { out } from "../../util/log";
-import { capture } from "../../util/notifications";
+import { OpenSearchConfigDirectAccess } from "./index";
+import { IndexFields } from "./index-based-on-resource";
 import {
-  BulkOperation,
   BulkResponseErrorItem,
   processErrorsFromBulkResponse,
   OnBulkItemError,
-} from "./bulk";
-import { OpenSearchConfigDirectAccess } from "./index";
-import { IndexFields } from "./index-based-on-resource";
-import { createLexicalDeleteQuery } from "./lexical/lexical-search";
+  resourceToBulkRequest,
+} from "./shared/bulk";
+import { createDeleteQuery } from "./shared/delete";
+import { getEntryId } from "./shared/id";
+import { getLog, OpenSearchLogLevel } from "./shared/log";
 
 dayjs.extend(duration);
 
@@ -34,7 +34,6 @@ const bulkChunkSize = 500;
 
 export type IngestRequest = IndexFields;
 export type IngestRequestResource = Omit<IndexFields, "cxId" | "patientId">;
-export type DeleteRequest = Pick<IndexFields, "cxId" | "patientId">;
 
 export type IngestBulkRequest = {
   cxId: string;
@@ -43,23 +42,26 @@ export type IngestBulkRequest = {
   onItemError?: OnBulkItemError | undefined;
 };
 
-export type OpenSearchFileIngestorDirectSettings = {
-  logLevel?: "info" | "debug" | "none";
+export type DeleteRequest = Pick<IndexFields, "cxId" | "patientId">;
+
+export type OpenSearchTextIngestorSettings = {
+  logLevel?: OpenSearchLogLevel;
 };
 
-export type OpenSearchTextIngestorDirectConfig = OpenSearchConfigDirectAccess & {
-  settings?: OpenSearchFileIngestorDirectSettings;
+export type OpenSearchTextIngestorConfig = OpenSearchConfigDirectAccess & {
+  settings?: OpenSearchTextIngestorSettings;
 };
 
-// TODO eng-41 Make this a factory so we can delegate the processing to a lambda
-export class OpenSearchTextIngestorDirect {
+/**
+ * Ingests text documents/entries in OpenSearch.
+ */
+export class OpenSearchTextIngestor {
   private readonly endpoint: string;
   private readonly username: string;
   private readonly password: string;
-  private readonly settings: OpenSearchFileIngestorDirectSettings;
+  private readonly settings: OpenSearchTextIngestorSettings;
 
-  constructor(readonly config: OpenSearchTextIngestorDirectConfig) {
-    // super(config);
+  constructor(readonly config: OpenSearchTextIngestorConfig) {
     this.endpoint = config.endpoint;
     this.username = config.username;
     this.password = config.password;
@@ -68,15 +70,15 @@ export class OpenSearchTextIngestorDirect {
     };
   }
 
-  async ingest({
+  async ingestSingle({
     cxId,
     patientId,
     resourceType,
     resourceId,
     content,
   }: IngestRequest): Promise<void> {
-    const defaultLogger = out(`ingest - ${cxId} - ${patientId}`);
-    const { log, debug } = this.getLog(defaultLogger);
+    const defaultLogger = out(`${this.constructor.name}.ingest - cx ${cxId}, pt ${patientId}`);
+    const { log, debug } = getLog(defaultLogger, this.settings.logLevel);
 
     const { indexName } = this.config;
     const auth = { username: this.username, password: this.password };
@@ -116,10 +118,16 @@ export class OpenSearchTextIngestorDirect {
    * @param resources - The resources to ingest.
    * @param onItemError - The function to call for each item error, optional. See buildOnItemError()
    *                      for the default implementation.
+   * @returns A map of error type as key and count of errors as value.
    */
-  async ingestBulk({ cxId, patientId, resources, onItemError }: IngestBulkRequest): Promise<void> {
-    const defaultLogger = out(`ingestBulk - ${cxId} - ${patientId}`);
-    const { log } = this.getLog(defaultLogger);
+  async ingestBulk({
+    cxId,
+    patientId,
+    resources,
+    onItemError,
+  }: IngestBulkRequest): Promise<Map<string, number>> {
+    const defaultLogger = out(`${this.constructor.name}.ingestBulk - cx ${cxId}, pt ${patientId}`);
+    const { log } = getLog(defaultLogger, this.settings.logLevel);
 
     const { indexName } = this.config;
     const auth = { username: this.username, password: this.password };
@@ -147,7 +155,7 @@ export class OpenSearchTextIngestorDirect {
 
     const time = Date.now() - startedAt;
     log(`Ingested ${resources.length} resources in ${time} ms, ${errorCount} errors`);
-    if (errors.size > 0) captureErrors({ cxId, patientId, errors, log });
+    return errors;
   }
 
   private async ingestBulkInternal({
@@ -161,8 +169,10 @@ export class OpenSearchTextIngestorDirect {
     indexName: string;
     client: Client;
   }): Promise<number> {
-    const defaultLogger = out(`...ingestBulkInternal - ${cxId} - ${patientId}`);
-    const { debug } = this.getLog(defaultLogger);
+    const defaultLogger = out(
+      `${this.constructor.name}.ingestBulkInternal - cx ${cxId}, pt ${patientId}`
+    );
+    const { debug } = getLog(defaultLogger, this.settings.logLevel);
 
     const operation = "index";
 
@@ -174,6 +184,7 @@ export class OpenSearchTextIngestorDirect {
         patientId,
         resource,
         operation,
+        getEntryId,
       })
     );
 
@@ -191,67 +202,24 @@ export class OpenSearchTextIngestorDirect {
   }
 
   async delete({ cxId, patientId }: DeleteRequest): Promise<void> {
-    const defaultLogger = out(`ingest - ${cxId} - ${patientId}`);
-    const { log, debug } = this.getLog(defaultLogger);
+    const defaultLogger = out(`${this.constructor.name}.delete - cx ${cxId}, pt ${patientId}`);
+    const { log, debug } = getLog(defaultLogger, this.settings.logLevel);
 
     const { indexName } = this.config;
-    const auth = { username: this.username, password: this.password };
-    const client = new Client({ node: this.endpoint, auth });
+    const auth = { username: this.config.username, password: this.config.password };
+    const client = new Client({ node: this.config.endpoint, auth });
 
     log(`Deleting resources from index ${indexName}...`);
     const startedAt = Date.now();
 
     const response = await client.deleteByQuery({
       index: indexName,
-      body: createLexicalDeleteQuery({ cxId, patientId }),
+      body: createDeleteQuery({ cxId, patientId }),
     });
     const time = Date.now() - startedAt;
     log(`Successfully deleted in ${time} milliseconds`);
     debug(`Response: `, () => JSON.stringify(response.body));
   }
-
-  private getLog(defaultLogger: ReturnType<typeof out>): ReturnType<typeof out> {
-    if (this.settings.logLevel === "none") return { debug: emptyFunction, log: emptyFunction };
-    return {
-      debug: this.settings.logLevel === "debug" ? defaultLogger.debug : emptyFunction,
-      log: defaultLogger.log,
-    };
-  }
-}
-
-function resourceToBulkRequest({
-  cxId,
-  patientId,
-  resource,
-  operation,
-}: {
-  cxId: string;
-  patientId: string;
-  resource: IngestRequestResource;
-  operation: BulkOperation;
-}) {
-  const entryId = getEntryId(cxId, patientId, resource.resourceId);
-  const document: IngestRequest = {
-    cxId,
-    patientId,
-    resourceType: resource.resourceType,
-    resourceId: resource.resourceId,
-    content: resource.content,
-  };
-  const cmd = { [operation]: { _id: entryId } };
-  return [cmd, document];
-}
-
-/**
- * Builds the ID for an OpenSearch entry.
- *
- * @param cxId - The cxId of the resource.
- * @param patientId - The patientId of the resource.
- * @param resourceId - The resourceId of the resource.
- * @returns The ID for the OpenSearch entry.
- */
-function getEntryId(cxId: string, patientId: string, resourceId: string): string {
-  return `${cxId}_${patientId}_${resourceId}`;
 }
 
 /**
@@ -267,22 +235,4 @@ function buildOnItemError(errors: Map<string, number>): OnBulkItemError {
     const count = errors.get(error.type) ?? 0;
     errors.set(error.type, count + 1);
   };
-}
-
-function captureErrors({
-  cxId,
-  patientId,
-  errors,
-  log,
-}: {
-  cxId: string;
-  patientId: string;
-  errors: Map<string, number>;
-  log: typeof console.log;
-}) {
-  const errorMapToObj = Object.fromEntries(errors.entries());
-  log(`Errors: `, () => JSON.stringify(errorMapToObj));
-  capture.error("Errors ingesting resources into OpenSearch", {
-    extra: { cxId, patientId, countPerErrorType: JSON.stringify(errorMapToObj, null, 2) },
-  });
 }
